requirements.txt
llama-index
llama-index-vector-stores-faiss
llama-index-embeddings-huggingface
llama-index-embeddings-clip
pymupdf
pdf2image
opencv-python-headless
faiss-cpu
sentence-transformers
torch
python-dotenv
ftfy 
regex 
tqdm
git+https://github.com/openai/CLIP.git
llama-index-llms-groq
langchain-text-splitters
langchain-community 
ragas 
datasets 
langchain-openai 
llama-index
langchain-groq

